feat #10: escribo un pequeño gestor de eventos que es lo suficientemente flexible para nuestras necesidades
