+++ app/src/main/java/com/project/agilugr/MainActivity.kt
@@ -9,6 +9,8 @@ import androidx.annotation.RequiresApi
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.core.view.MotionEventCompat
+import com.project.agilugr.gestures.DefaultGestureManager
+import com.project.agilugr.gestures.GestureManager
 import com.project.agilugr.ui.navigation.NavigationDirector
 import com.project.agilugr.ui.navigation.NavigationMapper
 import com.project.agilugr.ui.theme.AgilUGRTheme
@@ -23,6 +25,9 @@ class MainActivity : ComponentActivity() {
 
     // Tomamos el director de navegacion para que lance la interfaz grafica
     val navigation_director = NavigationDirector(focus_api = focus_api)
+
+    // Para gestionar todos los gestos de la aplicacion
+    val gestureManager = DefaultGestureManager(this.navigation_director)
 
     // Funcion principal
     @RequiresApi(Build.VERSION_CODES.O)
@@ -40,31 +45,11 @@ class MainActivity : ComponentActivity() {
         }
     }
 
-        // TODO -- mover a una gestion de eventos propia
-        override fun onTouchEvent(event: MotionEvent): Boolean {
-
-            val action: Int = MotionEventCompat.getActionMasked(event)
-
-            return when (action) {
-                MotionEvent.ACTION_DOWN -> {
-                    true
-                }
-                MotionEvent.ACTION_MOVE -> {
-                    true
-                }
-                MotionEvent.ACTION_UP -> {
-                    this.navigation_director.navigate(NavigationMapper.FOCUS_MODE_SESSION)
-                    true
-                }
-                MotionEvent.ACTION_CANCEL -> {
-                    true
-                }
-                MotionEvent.ACTION_OUTSIDE -> {
-                    true
-                }
-                else -> super.onTouchEvent(event)
-            }
-        }
+    /** Eventos tactiles sobre la pantalla del dispositivo */
+    override fun onTouchEvent(event: MotionEvent): Boolean {
+        // Empleamos el manager de gestos para procesar el evento
+        return this.gestureManager.onTouch(event)
+    }
 
 
 }
